extended taxonomy for RoPA
To have the fidesdemo be able to produce a RoPA compliant data map, we need to extend some resources. This will help users better understand the entire journey. These changes were made to try and realistically slot in as valid data uses for a valid data subject as part of this demo application. Coordination with the docs in the fides repo will still be required to walk through the extending of the taxonomy.